requirements.txt
autopep8==1.6.0
Faker==11.3.0
greenlet==1.1.2
psycopg2==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.29
text-unidecode==1.3
toml==0.10.2
